+++ src/main/java/com/buildstore/customer/repository/CustomerRepository.java
@@ -1,0 +1,7 @@
+package com.buildstore.customer.repository;
+
+import com.buildstore.customer.model.Customer;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+}