+++ src/main/java/com/buildstore/user/model/RoleName.java
@@ -1,0 +1,11 @@
+package com.buildstore.user.model;
+
+public enum RoleName {
+    ROLE_CUSTOMER,
+    ROLE_SALES_MANAGER,
+    ROLE_WAREHOUSE_MANAGER,
+    ROLE_PURCHASING_MANAGER,
+    ROLE_ACCOUNTANT,
+    ROLE_AUDITOR,
+    ROLE_ADMIN
+}